perf: decision matrices hold 0/1 in int8, not int64
PERF-7: every (rows x labels) decision matrix was int64 -- twice the
float32 probabilities it derives from, for an array of bits. At 250k rows
x 300 labels that is 600 MB per allocation, made and freed once per
threshold candidate, inside the phase the training memory budget exists
to protect. data.prepare_targets already refuses int64 for the target
matrix with the same arithmetic; these were missed.

sklearn scores an int8 indicator matrix identically, so the decisions are
unchanged -- pinned alongside the dtype.

+++ app/metrics.py
@@ -23,7 +23,7 @@ def argmax_onehot(proba: np.ndarray) -> np.ndarray:
     """One-hot argmax decision — the rule serving applies to binary/multiclass
     (``ClassifierModel.predict`` returns the single best label there and
     ignores thresholds entirely)."""
-    preds = np.zeros_like(proba, dtype=int)
+    preds = np.zeros_like(proba, dtype=np.int8)  # see apply_thresholds on the dtype
     preds[np.arange(proba.shape[0]), proba.argmax(axis=1)] = 1
     return preds
